refactor: remove manifest evaluation from policy engine
