+++ core/policy/policy-engine/build.gradle.kts
@@ -19,8 +19,6 @@ plugins {
 
 dependencies {
     api(project(":core:policy:policy-evaluator"))
-    api(project(":spi:policy-spi"))
-    api(project(":spi:transfer-spi"))
 }
 
 publishing {
