+++ core/policy/policy-evaluator/src/main/java/org/eclipse/dataspaceconnector/policy/model/AtomicConstraintFunction.java
@@ -12,11 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.policy.engine;
-
-import org.eclipse.dataspaceconnector.policy.model.AtomicConstraint;
-import org.eclipse.dataspaceconnector.policy.model.Operator;
-import org.eclipse.dataspaceconnector.policy.model.Rule;
+package org.eclipse.dataspaceconnector.policy.model;
 
 /**
  * An extension point that evaluates an {@link AtomicConstraint}.
