+++ spi/policy-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/PolicyEngine.java
@@ -9,7 +9,6 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
- *       Fraunhofer Institute for Software and Systems Engineering - resource manifest evaluation
  *
  */
 
@@ -21,13 +20,10 @@ import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.policy.model.Rule;
 import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
 import org.eclipse.dataspaceconnector.spi.policy.evaluation.AtomicConstraintFunction;
-import org.eclipse.dataspaceconnector.spi.policy.evaluation.ResourceDefinitionAtomicConstraintFunction;
-import org.eclipse.dataspaceconnector.spi.policy.evaluation.ResourceDefinitionRuleFunction;
 import org.eclipse.dataspaceconnector.spi.policy.evaluation.RuleFunction;
 import org.eclipse.dataspaceconnector.spi.result.Result;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceManifest;
 
+import java.util.Map;
 import java.util.function.BiFunction;
 
 /**
@@ -65,16 +61,10 @@ public interface PolicyEngine {
     Result<Policy> evaluate(String scope, Policy policy, ParticipantAgent agent);
     
     /**
-     * Evaluates and, if required, modifies the given ResourceManifest so that the given policy
-     * is fulfilled.
-     *
-     * @param scope the scope.
-     * @param policy the policy.
-     * @param resourceManifest the resource manifest to evaluate.
-     * @return a result containing either the verified/modified resource manifest or the problems
-     *         encountered during evaluation.
+     * Evaluates the given policy for an agent for the given scope using additional context information.
+     * Values in the map need to be of the same type defined by the key.
      */
-    Result<ResourceManifest> evaluate(String scope, Policy policy, ResourceManifest resourceManifest);
+    Result<Policy> evaluate(String scope, Policy policy, ParticipantAgent agent, Map<Class, Object> contextInformation);
 
     /**
      * Registers a function that is invoked when a policy contains an atomic constraint whose left operator expression evaluates to the given key for the specified scope.
@@ -94,28 +84,6 @@ public interface PolicyEngine {
      * @param function the function
      */
     <R extends Rule> void registerFunction(String scope, Class<R> type, RuleFunction<R> function);
-    
-    /**
-     * Registers a function that is invoked during resource manifest evaluation when a policy contains
-     * an atomic constraint whose left operator expression evaluates to the given key for the specified scope.
-     *
-     * @param scope the scope the function applies to.
-     * @param ruleType the {@link Rule} sub-type
-     * @param resourceDefinitionType the {@link ResourceDefinition} sub-type
-     * @param key the key.
-     * @param function the function.
-     */
-    <R extends Rule, D extends ResourceDefinition> void registerFunction(String scope, Class<R> ruleType, Class<D> resourceDefinitionType, String key, ResourceDefinitionAtomicConstraintFunction<R, D> function);
-    
-    /**
-     * Registers a function that is invoked during resource manifest evaluation when a policy contains a rule of the given type for the specified scope.
-     *
-     * @param scope the scope the function applies to.
-     * @param ruleType the {@link Rule} sub-type
-     * @param resourceDefinitionType the {@link ResourceDefinition} sub-type
-     * @param function the function
-     */
-    <R extends Rule, D extends ResourceDefinition> void registerFunction(String scope, Class<R> ruleType, Class<D> resourceDefinitionType, ResourceDefinitionRuleFunction<R, D> function);
 
     /**
      * Registers a function that performs pre-validation on the policy for the given scope.
